Conditional render inside return. Potential needless code
Created render of single year or over time as separate renders based on whether viewMode is single or trend. Did literally this in Haskell earlier.

Commented out the if showExported, if showRenewables, as I think this can be done in a different way. Currently, too much data is being passed to the single functions which shouldn't render all the data in a single call maybe. Multiple graphs in a file or done seperately?