Graph accepts different data
Changed the graph so exported and renewables shows up. changed the single year choose data useEffect so that it creates the data based on the new useState: selectedEnergyType. Modified radiobuttons so they only modify a single state now, and input a text value.

Error with rendering the data at all when clicking radiobuttons, but it works manually inputting the name into the useState and the dependency array is updated.